+++ src/lib.rs
@@ -66,6 +66,43 @@
 //! # }
 //! ```
 //!
+//! Checking if the "primary" clipboard is supported (note that this might be unnecessary depending
+//! on your crate usage, the regular copying and pasting functions do report if the primary
+//! selection is unsupported when it is requested):
+//!
+//! ```no_run
+//! # extern crate wl_clipboard_rs;
+//! # extern crate failure;
+//! # use failure::Error;
+//! # fn foo() -> Result<(), Error> {
+//! use wl_clipboard_rs::utils::{is_primary_selection_supported, PrimarySelectionCheckError};
+//!
+//! match is_primary_selection_supported() {
+//!     Ok(supported) => {
+//!         // We have our definitive result. This means that either data-control version 1 is
+//!         // present (which does not support the primary selection), or that data-control
+//!         // version 2 is present and it did not signal the primary selection support.
+//!     },
+//!     Err(PrimarySelectionCheckError::NoSeats) => {
+//!         // Impossible to give a definitive result. Primary selection may or may not be
+//!         // supported.
+//!
+//!         // The required protocol (data-control version 2) is there, but there are no seats.
+//!         // Unfortunately, at least one seat is needed to check for the primary clipboard
+//!         // support.
+//!     },
+//!     Err(PrimarySelectionCheckError::MissingProtocol { .. }) => {
+//!         // The data-control protocol (required for wl-clipboard-rs operation) is not
+//!         // supported by the compositor.
+//!     },
+//!     Err(_) => {
+//!         // Some communication error occurred.
+//!     }
+//! }
+//! # Ok(())
+//! # }
+//! ```
+//!
 //! # Included terminal utilities
 //!
 //! - `wl-paste`: implements `wl-paste` from