+++ src/utils.rs
@@ -1,6 +1,13 @@
 //! Helper functions.
 
-use std::{ffi::CString, os::unix::io::RawFd, process::abort};
+use std::{
+    cell::{Cell, RefCell},
+    ffi::{CString, OsString},
+    io,
+    os::unix::io::RawFd,
+    process::abort,
+    rc::Rc,
+};
 
 use failure::Fail;
 use libc::{STDIN_FILENO, STDOUT_FILENO};
@@ -8,6 +15,16 @@ use nix::{
     fcntl::{fcntl, FcntlArg, OFlag},
     sys::wait::{waitpid, WaitStatus},
     unistd::{close, dup2, execvp, fork, ForkResult},
+};
+use wayland_client::{
+    global_filter, protocol::wl_seat::WlSeat, ConnectError, Display, GlobalError, GlobalManager,
+    Interface, NewProxy,
+};
+use wayland_protocols::wlr::unstable::data_control::v1::client::zwlr_data_control_manager_v1::ZwlrDataControlManagerV1;
+
+use crate::{
+    handlers::{DataDeviceHandler, WlSeatHandler},
+    seat_data::SeatData,
 };
 
 /// Checks if the given MIME type represents plain text.
@@ -156,4 +173,124 @@ pub fn copy_data(from_fd: Option<RawFd>, to_fd: RawFd, wait: bool) -> Result<(),
     }
 
     Ok(())
+}
+
+/// Errors that can occur when checking whether the primary selection is supported.
+#[derive(Fail, Debug)]
+pub enum PrimarySelectionCheckError {
+    #[fail(display = "There are no seats")]
+    NoSeats,
+
+    #[fail(display = "Couldn't connect to the Wayland compositor")]
+    WaylandConnection(#[cause] ConnectError),
+
+    #[fail(display = "Wayland compositor communication error")]
+    WaylandCommunication(#[cause] io::Error),
+
+    #[fail(display = "A required Wayland protocol ({} version {}) is not supported by the compositor",
+           name, version)]
+    MissingProtocol { name: &'static str, version: u32 },
+}
+
+/// Checks if the compositor supports the primary selection.
+///
+/// # Examples
+///
+/// ```no_run
+/// # extern crate wl_clipboard_rs;
+/// # extern crate failure;
+/// # use failure::Error;
+/// # fn foo() -> Result<(), Error> {
+/// use wl_clipboard_rs::utils::{is_primary_selection_supported, PrimarySelectionCheckError};
+///
+/// match is_primary_selection_supported() {
+///     Ok(supported) => {
+///         // We have our definitive result. This means that either data-control version 1 is
+///         // present (which does not support the primary selection), or that data-control
+///         // version 2 is present and it did not signal the primary selection support.
+///     },
+///     Err(PrimarySelectionCheckError::NoSeats) => {
+///         // Impossible to give a definitive result. Primary selection may or may not be
+///         // supported.
+///
+///         // The required protocol (data-control version 2) is there, but there are no seats.
+///         // Unfortunately, at least one seat is needed to check for the primary clipboard
+///         // support.
+///     },
+///     Err(PrimarySelectionCheckError::MissingProtocol { .. }) => {
+///         // The data-control protocol (required for wl-clipboard-rs operation) is not
+///         // supported by the compositor.
+///     },
+///     Err(_) => {
+///         // Some communication error occurred.
+///     }
+/// }
+/// # Ok(())
+/// # }
+/// ```
+#[inline]
+pub fn is_primary_selection_supported() -> Result<bool, PrimarySelectionCheckError> {
+    is_primary_selection_supported_internal(None)
+}
+
+pub(crate) fn is_primary_selection_supported_internal(
+    socket_name: Option<OsString>)
+    -> Result<bool, PrimarySelectionCheckError> {
+    // Connect to the Wayland compositor.
+    let (display, mut queue) = match socket_name {
+                                   Some(name) => Display::connect_to_name(name),
+                                   None => Display::connect_to_env(),
+                               }.map_err(PrimarySelectionCheckError::WaylandConnection)?;
+
+    let seats = Rc::new(RefCell::new(Vec::<WlSeat>::new()));
+
+    let seats_2 = seats.clone();
+    let global_manager =
+        GlobalManager::new_with_cb(&display,
+                                   global_filter!([WlSeat, 6, move |seat: NewProxy<WlSeat>| {
+                                                      let seat_data =
+                                                          RefCell::new(SeatData::default());
+                                                      let seat =
+                                                          seat.implement(WlSeatHandler, seat_data);
+                                                      seats_2.borrow_mut().push(seat.clone());
+                                                      seat
+                                                  }]));
+
+    // Retrieve the global interfaces.
+    queue.sync_roundtrip()
+         .map_err(PrimarySelectionCheckError::WaylandCommunication)?;
+
+    // Try instantiating data control version 2. If data control is missing altogether, return a
+    // missing protocol error, but if it's present with version 1 then return false as version 1
+    // does not support primary clipboard.
+    let impl_manager = |manager: NewProxy<_>| manager.implement_dummy();
+    let clipboard_manager =
+        match global_manager.instantiate_exact::<ZwlrDataControlManagerV1, _>(2, impl_manager) {
+            Ok(manager) => manager,
+            Err(GlobalError::Missing) => {
+                return Err(PrimarySelectionCheckError::MissingProtocol { name: ZwlrDataControlManagerV1::NAME,
+                                                     version: 1 })
+            }
+            Err(GlobalError::VersionTooLow(_)) => return Ok(false),
+        };
+
+    // Check if there are no seats.
+    if seats.borrow_mut().is_empty() {
+        return Err(PrimarySelectionCheckError::NoSeats);
+    }
+
+    let supports_primary = Rc::new(Cell::new(false));
+
+    // Go through the seats and get their data devices. They will listen for the primary_selection
+    // event and set supports_primary on receiving one.
+    for seat in &*seats.borrow_mut() {
+        let handler = DataDeviceHandler::new(seat.clone(), true, supports_primary.clone());
+        clipboard_manager.get_data_device(seat, |device| device.implement(handler, ()))
+                         .unwrap();
+    }
+
+    queue.sync_roundtrip()
+         .map_err(PrimarySelectionCheckError::WaylandCommunication)?;
+
+    Ok(supports_primary.get())
 }